Merged administrators.
Fixed the send EMail to the administrator so that it now retrieves the From Address from
the users table in the amavis database (getFromMAilAddr). For this update to work properly
the users table needs to track the external authentication system(s).

Fixed invocation of the mailer class so that it does not throw external exceptions.