Merged administrators.
Renamed the function s_admin to isAdmin and updated the code to check the merged administrators.

Removed the function isMailAdmin. Fixed the COOKIE login process to recover missing data from
the users table. It used to set all variables to the COOKIE_ID.

Added a cookie timeout configuration variable.

Fixed some translations and added some where there were none.

Added synchronisaton code for the users database. When a user logs in the users table entries
are brought in line with the data that came from the authentication system.